+++ eras/alonzo/impl/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.1.0
 
 * Add `ToJSON` instance for `AlonzoTxOut`, `AlonzoScript` and `Datum`
+* Add `ToJSON` instance for `AlonzoPParams StrictMaybe`
 
 ## 1.0.0.0