+++ eras/babbage/impl/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 1.0.1.0
 
-* Add `ToJSON` instance for `BabbageTxOut`
+* Add `ToJSON` instance for `BabbageTxOut`.
+* Add `ToJSON` instance for `BabbagePParams Identity` and `BabbagePParams StrictMaybe`
 
 ## 1.0.0.0